applet.interface.spi_controller: migrate to V2 API.
This also unfortunately removes functionality (only Mode 3 is supported
in the new controller, whereas the old one supported all modes), but
most devices use either Mode 0 or Mode 3 and don't care which one it is.
