+++ docs/manual/src/applets/interface/spi_controller.rst
@@ -1,0 +1,7 @@
+``spi-controller``
+==================
+
+.. _applet.interface.spi_controller:
+
+.. autoprogram:: glasgow.applet.interface.spi_controller:SPIControllerApplet._get_argparser_for_sphinx("spi-controller")
+    :prog: glasgow run spi-controller
