+++ software/glasgow/applet/display/pdi/__init__.py
@@ -65,7 +65,7 @@ from glasgow.support.bits import bitarray
 from amaranth import *
 from amaranth.lib import io
 
-from ...interface.spi_controller import SPIControllerSubtarget, SPIControllerInterface
+from ...interface.spi_controller_deprecated import SPIControllerSubtarget, SPIControllerInterface
 from ... import *
